Move upgrades into app package to avoid import cycles
